Remove old webpack-based service worker
