+++ portal/pnpm-workspace.yaml
@@ -2,4 +2,3 @@ packages:
     - 'common'
     - 'worker'
     - 'server'
-    - 'worker-nextjs'
