--- portal/worker-nextjs/package.json
@@ -1,28 +1,0 @@
-{
-	"name": "worker",
-	"dependencies": {
-		"@mysten/sui": "^1.12.0",
-		"common": "workspace:common",
-		"next": "^15.0.4",
-		"react": "^18.3.1",
-		"ts-loader": "^9.5.1",
-		"tsc": "^2.0.4",
-		"typescript": "^5.6.3"
-	},
-	"devDependencies": {
-		"@types/serviceworker": "^0.0.107",
-		"copy-webpack-plugin": "^12.0.2",
-		"css-minimizer-webpack-plugin": "^7.0.0",
-		"html-minimizer-webpack-plugin": "^5.0.0",
-		"vitest": "^2.1.3",
-		"webpack": "^5.95.0",
-		"webpack-cli": "^5.1.4",
-		"webpack-dev-server": "^5.1.0",
-		"webpack-merge": "^6.0.1"
-	},
-	"scripts": {
-		"dev": "next dev",
-		"dev:prod": "next prod",
-		"build": "webpack --no-watch --config webpack.config.prod.js"
-	}
-}
