--- portal/worker-nextjs/tsconfig.json
@@ -1,55 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ESNext",
-    "lib": [
-      "dom",
-      "dom.iterable",
-      "esnext"
-    ],
-    "allowJs": true,
-    "skipLibCheck": true,
-    "strict": false,
-    "forceConsistentCasingInFileNames": true,
-    "noEmit": true,
-    "incremental": true,
-    "esModuleInterop": true,
-    "module": "esnext",
-    "moduleResolution": "node",
-    "resolveJsonModule": true,
-    "isolatedModules": true,
-    "jsx": "preserve",
-    "composite": true,
-    "baseUrl": "./",
-    "paths": {
-      "@static/*": [
-        "../common/static/*"
-      ],
-      "@lib/*": [
-        "../common/lib/*"
-      ]
-    },
-    "plugins": [
-      {
-        "name": "next"
-      }
-    ]
-  },
-  "baseUrl": "./",
-  "rootDir": "src",
-  "outDir": "dist",
-  "skipLibCheck": true,
-  "include": [
-    "../common/**/*",
-    "src/**/*",
-    ".next/types/**/*.ts"
-  ],
-  "exclude": [
-    "node_modules",
-    "dist"
-  ],
-  "references": [
-    {
-      "path": "../common/tsconfig.json"
-    }
-  ]
-}
