--- portal/worker/concrete.css
@@ -1,6 +1,0 @@
-/*
-  Copyright (c) Mysten Labs, Inc.
-  SPDX-License-Identifier: Apache-2.0
-*/
-
-/*! concrete.css v2.1.1 | MIT License | github.com/louismerlin/concrete.css */:root{--fg:#111;--bg:#fff}html{box-sizing:border-box;font-size:62.5%}*,:after,:before{box-sizing:inherit;text-decoration-thickness:.2rem}body{background:var(--bg);font-family:Helvetica,Arial,sans-serif;font-size:2rem;font-weight:400}a,body{color:var(--fg)}img{height:auto;max-width:100%}main{margin:auto;min-height:100vh;padding:0 80px;width:100%}header{font-size:1.5em;padding:3rem 0 0}.inverted{background-color:#030f1c;color:var(--bg);padding:10px}footer{align-items:center;border:0;border-top:1px solid hsla(0,0%,41%,.3);color:#696969;display:flex;flex-direction:row;font-size:14px;justify-content:space-between;letter-spacing:.14px;line-height:100%;padding:30px 0;width:100%}@media (max-width:768px){footer{flex-direction:column;gap:10px;padding:20px 0}main{padding:0 40px}}section{padding:4rem 0}button,input[type=button],input[type=reset],input[type=submit]{background:var(--bg);border:.2rem solid var(--fg);border-radius:0;color:var(--fg);cursor:pointer;display:inline-block;font-size:2rem;font-weight:400;margin-bottom:1rem;padding:.4rem 1rem;vertical-align:middle}button:disabled,input[type=button]:disabled,input[type=reset]:disabled,input[type=submit]:disabled{border-style:dashed;cursor:not-allowed}ul{list-style:square}fieldset{border:.2rem solid var(--fg)}label,legend{display:block;font-weight:700;margin-bottom:.8rem}input[type=email],input[type=number],input[type=password],input[type=search],input[type=tel],input[type=text],input[type=url],select,textarea{-webkit-appearance:none;-moz-appearance:none;appearance:none;background-color:var(--bg);border:.2rem solid var(--fg);border-radius:0;box-shadow:none;box-sizing:inherit;color:var(--fg);font-size:2rem;padding:.4rem 1rem;width:100%}fieldset,input,select,textarea{margin:0 0 1.6rem}input::placeholder,textarea::placeholder{color:var(--fg);font-style:italic}table{border-spacing:0;width:100%}td,th{padding:.8rem}td:first-child,th:first-child{padding-left:0}td:last-child,th:last-child{padding-right:0}th{border-bottom:.2rem solid var(--fg);text-align:left}td{border-bottom:.1rem solid var(--fg)}blockquote,pre{border-left:.2rem solid var(--fg);margin-left:0;margin-right:0;overflow-y:hidden;padding:1rem 1.6rem}pre{border:.1rem dotted var(--fg);border-left:.2rem solid var(--fg)}pre>code{display:block;font-size:1.6rem;white-space:pre}progress{-moz-appearance:none;-webkit-appearance:none;background:var(--bg);border:.2rem solid var(--fg);border-radius:0;color:var(--fg);display:block;height:1rem;overflow:hidden;padding:0;width:100%}progress::-webkit-progress-bar{background-color:var(--bg)}progress::-webkit-progress-value{background-color:var(--fg)}progress::-moz-progress-bar{background-color:var(--fg)}hr{border:.1rem solid var(--fg)}@font-face{font-family:AlumniSans;src:url(/fonts/AlumniSans-Bold.ttf) format("truetype")}.AlumniSans{font-family:AlumniSans,sans-serif}@font-face{font-family:InterTightMedium;src:url(/fonts/InterTight-Medium.ttf) format("truetype")}.InterTightMedium{font-family:InterTightMedium,sans-serif}@font-face{font-family:InterTightBold;src:url(/fonts/InterTight-Bold.ttf) format("truetype")}.InterTightBold{font-family:InterTightBold,sans-serif}.content-div{flex:1 1 0%;margin-top:40px;max-width:600px;row-gap:40px}.content-div,.main-div{align-items:center;display:flex;flex-direction:column;justify-content:center}.main-div{color:#000;gap:40px;min-height:100vh;width:100%}.docs{align-items:center;background-color:#ee1a32;color:#fff;display:inline-flex;flex-direction:column;font-size:16px;justify-content:center;line-height:160%;padding:12px 40px;text-decoration:none;width:auto}.start{align-items:start}.center,.start{display:flex;flex-direction:column;justify-content:center;width:100%}.center{align-items:center}.spinner{animation:spin 1s linear infinite;border:3px solid rgba(0,0,0,.1);border-radius:50%;border-top-color:#030f1c;height:24px;margin-bottom:12px;width:24px}@keyframes spin{to{transform:rotate(1turn)}}.custom-header{color:#000;font-size:90px;font-style:normal;letter-spacing:.96px;line-height:100%;text-align:left;width:100%}@media (max-width:768px){.custom-header{font-size:70px;line-height:90%}}.custom-p{color:#696969;font-size:16px;font-style:normal;font-weight:500;line-height:160%}.terms-policies{align-items:center;display:inline-flex;gap:32px;justify-content:center}.collapsible-content{display:none}.collapsible-content.show{display:flex}.learn-more{color:blue;cursor:pointer;text-decoration:underline}.spinner-text{color:#696969;font-size:16px;font-style:normal;font-weight:500;line-height:160%}
