+++ portal/worker/src/register.ts
